fixed cart

+++ react-app/src/components/CartItem.js
@@ -7,6 +7,7 @@ const CartItem = ({ item, cart, product, handleDelete }) => {
   const dispatch = useDispatch()
 
   useEffect(() => {
+    console.log("!!!!QUANTITY!!!!", quantity)
     dispatch(updateShoppingCart(item?.id, quantity, cart?.id))
   }, [dispatch, quantity])
 
@@ -29,7 +30,8 @@ const CartItem = ({ item, cart, product, handleDelete }) => {
               type="button"
               onClick={(e) => {
                 e.preventDefault();
-                handleDelete(product);
+                const confirmed = window.confirm(`remove ${product.name} from your cart?`)
+                if(confirmed) handleDelete(product);
               }}
             >
               <i className="far fa-trash-alt"></i>
@@ -49,6 +51,7 @@ const CartItem = ({ item, cart, product, handleDelete }) => {
               <button
                 type="button"
                 onClick={(e) => {
+                  e.preventDefault()
                   setQuantity(quantity + 1);
                 }}
               >
@@ -58,6 +61,7 @@ const CartItem = ({ item, cart, product, handleDelete }) => {
               <button
                 type="button"
                 onClick={(e) => {
+                  e.preventDefault()
                   setQuantity(quantity - 1);
                   if (quantity <= 1) setQuantity(1);
                 }}
